Adds event handling for mouse clicks by user.
Will now detect mouse clicks made by the user and correctly assign these clicks to the correct square on the board.
Does not print the pieces on the board just yet (this is the last thing to implement before the AI can be worked on).

There is also an issue with global variables being defined in every file separately. This is ugly and they should only be assigned values in one place.
Will try to find a fix for this later.